+++ tsfile/src/main/java/org/apache/iotdb/tsfile/encoding/decoder/LongGorillaDecoder.java
@@ -66,7 +66,7 @@ public class LongGorillaDecoder extends GorillaDecoderV2 {
 
   @SuppressWarnings("squid:S128")
   protected long readNext(ByteBuffer in) {
-    int controlBits = readNextClearBit(2, in);
+    byte controlBits = readNextClearBit(2, in);
 
     switch (controlBits) {
       case 3: // case '11': use new leading and trailing zeros
